test: Shift serde tests to 'toml'
Depending on the case, a serde tests could live in either or both
crates, so let's simplify by just saying all serde tests live in `toml`.
This as the added benefit in that it makes sure the newtype wrappers
don't break things.
