+++ crates/toml_edit/tests/testsuite/main.rs
@@ -2,13 +2,6 @@
 
 mod convert;
 mod edit;
-mod enum_external_deserialize;
-mod formatter;
 mod invalid;
-mod macros;
 mod parse;
-mod pretty;
-mod serde;
-mod spanned;
-mod spanned_impls;
 mod stackoverflow;
